+++ config.yaml
@@ -15,7 +15,7 @@ mysql:
   host: "127.0.0.1"
   port: 3306
   user: "root"
-  password: "123456"
+  password: "13572468"
   dbname: "lmp"
   max_open_conns: 200
   max_idle_conns: 50
