+++ plugins/Test/sc01.py
@@ -1,0 +1,4 @@
+import utils
+
+if __name__ == '__main__':
+    pass
