+++ plugins/c/PickNext.c
@@ -1,0 +1,49 @@
+#include <uapi/linux/ptrace.h>
+
+struct key_t {
+    u32 cpu;
+    u32 pid;
+    u32 tgid;
+};
+
+BPF_HASH(start, struct key_t);
+BPF_HASH(dist, struct key_t);
+
+int pick_start(struct pt_regs *ctx)
+{
+    u64 ts = bpf_ktime_get_ns();
+    u64 pid_tgid = bpf_get_current_pid_tgid();
+    struct key_t key;
+
+    key.cpu = bpf_get_smp_processor_id();
+    key.pid = pid_tgid;
+    key.tgid = pid_tgid >> 32;
+
+    start.update(&key, &ts);
+    return 0;
+}
+
+int pick_end(struct pt_regs *ctx)
+{
+    u64 ts = bpf_ktime_get_ns();
+    u64 pid_tgid = bpf_get_current_pid_tgid();
+    struct key_t key;
+    u64 *value;
+    u64 delta;
+
+    key.cpu = bpf_get_smp_processor_id();
+    key.pid = pid_tgid;
+    key.tgid = pid_tgid >> 32;
+
+    value = start.lookup(&key);
+
+    if (value == 0) {
+        return 0;
+    }
+
+    delta = ts - *value;
+    start.delete(&key);
+    dist.increment(key, delta);
+
+    return 0;
+}
