+++ plugins/c/vfsstat.c
@@ -1,0 +1,23 @@
+#include <uapi/linux/ptrace.h>
+
+enum stat_types {
+    S_READ = 1,
+    S_WRITE,
+    S_FSYNC,
+    S_OPEN,
+    S_CREATE,
+    S_MAXSTAT
+};
+
+BPF_ARRAY(stats, u64, S_MAXSTAT);
+
+static void stats_increment(int key) {
+    u64 *leaf = stats.lookup(&key);
+    if (leaf) (*leaf)++;
+}
+
+void do_read(struct pt_regs *ctx) { stats_increment(S_READ); }
+void do_write(struct pt_regs *ctx) { stats_increment(S_WRITE); }
+void do_fsync(struct pt_regs *ctx) { stats_increment(S_FSYNC); }
+void do_open(struct pt_regs *ctx) { stats_increment(S_OPEN); }
+void do_create(struct pt_regs *ctx) { stats_increment(S_CREATE); }
