+++ plugins/settings/config.py
@@ -4,13 +4,14 @@
 import yaml
 import os
 
-current_path = os.path.abspath(".")
+current_path = os.path.abspath("..")
 yaml_path = os.path.join(current_path, "config.yaml")
 
 
 def read_config():
-    with open(yaml_path,'r') as stream:
-        cfg =yaml.load(stream,Loader=yaml.FullLoader)
+    with open(yaml_path, 'r') as stream:
+        cfg = yaml.load(stream, Loader=yaml.FullLoader)
     return cfg
+
 
 cfg = read_config()
