+++ plugins/settings/const.py
@@ -1,5 +1,6 @@
 from enum import IntEnum
 
+
 class DatabaseType(IntEnum):
     INFLUXDB = 1
     ES = 2
